+++ controllers/postController.js
@@ -74,7 +74,6 @@ exports.getPost = asyncHandler(async (req, res, next) => {
     }
 });
 
-// TODO: Verify that the the posts user matches the id supplied by the token (On both update and delete)
 exports.updatePost = [
     body("title", "Posts must include a title")
         .trim()
@@ -88,39 +87,62 @@ exports.updatePost = [
     asyncHandler(async (req, res, next) => {
         const errors = validationResult(req);
 
-        if (!errors.isEmpty()) {
-            // Inform client post had errors
-            res.status(400).json({
-                post: { title: req.body.title, content: req.body.content },
-                errors: errors.array(),
-            });
-            return;
-        } else {
-            const post = await Post.findByIdAndUpdate(req.params.postId, {
-                title: req.body.title,
-                content: req.body.content,
-            });
+        const post = await Post.findOne({ _id: req.params.postId }, "user")
+            .populate("user", { name: 1 })
+            .exec();
 
-            if (!post) {
-                return res.status(404).json({
-                    error: `No post with id ${req.params.postId} exists`,
+        if (!post) {
+            return res.status(404).json({
+                error: `No post with id ${req.params.postId} exists`,
+            });
+        }
+
+        if (post.user._id === req.user._id) {
+            if (!errors.isEmpty()) {
+                // Inform client post had errors
+                res.status(400).json({
+                    post: { title: req.body.title, content: req.body.content },
+                    errors: errors.array(),
                 });
+                return;
             } else {
-                res.json({ message: "Post updated successfully", post: post });
+                const post = await Post.findByIdAndUpdate(req.params.postId, {
+                    title: req.body.title,
+                    content: req.body.content,
+                });
+
+                res.json({
+                    message: "Post updated successfully",
+                    post: post,
+                });
             }
+        } else {
+            res.status(401).json({ error: "Not authorized for this action" });
         }
     }),
 ];
 
 exports.deletePost = asyncHandler(async (req, res, next) => {
-    const post = await Post.findByIdAndDelete(req.params.postId);
+    const post = await Post.findOne({ _id: req.params.postId }, "user")
+        .populate("user", { name: 1 })
+        .exec();
 
     if (!post) {
         return res
             .status(404)
             .json({ error: `No post with id ${req.params.postId} exists` });
-    } else {
-        const comment = await Comment.deleteMany({ postId: req.params.postId });
+    }
+
+    if (post.user._id === req.user._id) {
+        const post = await Post.findByIdAndDelete(req.params.postId);
+        await Comment.deleteMany({
+            postId: req.params.postId,
+        });
+        await User.findByIdAndUpdate(req.user._id, {
+            $pull: { posts: req.params.postId },
+        });
         res.json({ message: "Post deleted successfully", post: post });
+    } else {
+        res.status(401).json({ error: "Not authorized for this action" });
     }
 });
